+++ src/main/java/net/Gabou/projectatmosphere/mixin/client/DefaultPipelineTornadoMixin.java
@@ -33,9 +33,11 @@ public abstract class DefaultPipelineTornadoMixin {
         float[] cloudColor = renderer.getCloudColor(partialTick);
         mc.getProfiler().push("projectatmosphere_tornado_opaque");
         SimpleCloudsTornadoRenderer.INSTANCE.prepareFrame(level, partialTick);
+        renderer.copyDepthFromCloudsToTransparency();
         renderer.getCloudTarget().bindWrite(false);
         SimpleCloudsTornadoRenderer.INSTANCE.renderOpaque(
-                renderer, stack, projMat, partialTick, cloudColor[0], cloudColor[1], cloudColor[2]
+                renderer, stack, projMat, partialTick, cloudColor[0], cloudColor[1], cloudColor[2],
+                renderer.getCloudTransparencyTarget().getDepthTextureId(), true
         );
         mc.getProfiler().pop();
     }